+++ heimdallsword/data/config.py
@@ -25,13 +25,13 @@ class Config:
 
     DEFAULT_SENDERS_FILE = "senders.txt"
     """ Senders file default name """
-    
+
     DEFAULT_RECIPIENTS_FILE = "recipients.txt"
     """ Recipients file default name """
-    
+
     DEFAULT_CONTENT_DIR  = "content"
     """ Content directory default name """
-    
+
     DEFAULT_DELAY = 100
     """ Default delay value in milliseconds used between each email sent """
 
@@ -40,18 +40,19 @@ class Config:
     Default metrics delay value in seconds used to wait after sending an email
     before logging into the sender's account and retrieving bounced emails
     """
-    
+
     DEFAULT_SMTP_PORT = 587
     """ Default SMTP port """
-    
+
     DEFAULT_POP3_PORT = 995
     """ Default POP3 SSL port """
-    
+
     DEFAULT_LOG_FILE_PATH = "./heimdallsword.log"
     """ Default log file name and path """
-    
+
     DEFAULT_METRICS_FILE_PATH = "./metrics.txt"
     """ Default metrics file name and path """
+
 
     def __init__(self):
         self.recipient_file = ""