+++ heimdallsword/release.py
@@ -12,13 +12,13 @@ This module contains release information about this package.
 
 # Package imformation
 __package__     = "heimdallsword"
-__version__     = "0.0.1"
+__version__     = "0.0.2"
 __license__     = "BSD-3-Clause"
 __author__      = "eldiablo"
 __email__       = "avsarria@gmail.com"
 __url__         = "http://github.com/rwprimitives/heimdallsword"
 __bug_tracker__ = "https://github.com/rwprimitives/heimdallsword/issues"
-__date__        = "06/26/2022"
+__date__        = "06/29/2022"
 __copyright__   = "Copyright (c) 2022 rwprimitives"
 __description__ = "A tool for sending emails to multiple recipients using " \
                   "one or multiple sender accounts with customizable email "  \