+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 
 ****
 
+[0.0.2] - 2022-06-29
+--------------------
+- Fixed all flake8 warnings
+- Fixed bug in renderer module. An exception is thrown when not enough space is available in terminal window to draw windows
+- Added new exit code to indicate when the program fails due to not enough space available in terminal window for graphical panel
+
 [0.0.1] - 2022-06-26
 --------------------
 - Initial release